test: extract prompt cache white-box suites
